+++ src/pages/shared_modules/styles.py
